Revert "Feat: Single Source Read (#573)"
This reverts commit d1027d51e3f08531a1c2ea1ff58621e5c7f279ef.
