--- packages/amplify-category-api/src/graphql-transformer/constants.ts
@@ -1,6 +1,0 @@
-export const DESTRUCTIVE_UPDATES_FLAG = 'allow-destructive-graphql-schema-updates';
-export const PROVIDER_NAME = 'awscloudformation';
-export const PARAMETERS_FILENAME = 'parameters.json';
-export const ROOT_APPSYNC_S3_KEY = 'amplify-appsync-files';
-export const SCHEMA_FILENAME = 'schema.graphql';
-export const SCHEMA_DIR_NAME = 'schema';
