+++ packages/amplify-category-api/src/graphql-transformer/transform-graphql-schema-v2.ts
@@ -2,23 +2,34 @@ import {
   collectDirectivesByTypeNames,
   DeploymentResources,
   GraphQLTransform,
+  ResolverConfig,
+  TransformerProjectConfig,
 } from '@aws-amplify/graphql-transformer-core';
-import { AppSyncAuthConfiguration } from '@aws-amplify/graphql-transformer-interfaces';
+import { Template } from '@aws-amplify/graphql-transformer-core/lib/config/project-config';
+import { OverrideConfig } from '@aws-amplify/graphql-transformer-core/lib/transformation/types';
+import { AppSyncAuthConfiguration, TransformerPluginProvider } from '@aws-amplify/graphql-transformer-interfaces';
 import {
+  $TSAny,
   $TSContext,
+  $TSMeta,
   $TSObject,
   AmplifyCategories,
   AmplifySupportedService,
   ApiCategoryFacade,
+  CloudformationProviderFacade,
   getGraphQLTransformerAuthDocLink,
   JSONUtilities,
   pathManager,
+  stateManager,
 } from 'amplify-cli-core';
 import { printer } from 'amplify-prompts';
 import fs from 'fs-extra';
 import { ResourceConstants } from 'graphql-transformer-common';
 import {
+  DiffRule,
+  getSanityCheckRules,
   loadProject,
+  ProjectRule,
   sanityCheckProject,
 } from 'graphql-transformer-core';
 import _ from 'lodash';
@@ -27,29 +38,24 @@ import path from 'path';
 import { searchablePushChecks } from './api-utils';
 import { AmplifyCLIFeatureFlagAdapter } from './amplify-cli-feature-flag-adapter';
 import { isAuthModeUpdated } from './auth-mode-compare';
-import {
-  schemaHasSandboxModeEnabled,
-  showGlobalSandboxModeWarning,
-  showSandboxModePrompts,
-} from './sandbox-mode-helpers';
+import { schemaHasSandboxModeEnabled, showGlobalSandboxModeWarning, showSandboxModePrompts } from './sandbox-mode-helpers';
 import { parseUserDefinedSlots } from './user-defined-slots';
 import {
-  mergeUserConfigWithTransformOutput,
-  writeDeploymentToDisk,
+  getAdminRoles, getIdentityPoolId, mergeUserConfigWithTransformOutput, writeDeploymentToDisk,
 } from './utils';
-import {
-  generateTransformerOptions,
-} from './transformer-options-v2';
-import {
-  PARAMETERS_FILENAME,
-  SCHEMA_DIR_NAME,
-  SCHEMA_FILENAME,
-} from './constants';
-import {
-  TransformerFactoryArgs,
-  TransformerProjectOptions,
-} from './transformer-options-types';
-import { contextUtil } from '../category-utils/context-util';
+import { getTransformerFactory } from './transformer-factory';
+
+const PARAMETERS_FILENAME = 'parameters.json';
+const SCHEMA_FILENAME = 'schema.graphql';
+const SCHEMA_DIR_NAME = 'schema';
+const ROOT_APPSYNC_S3_KEY = 'amplify-appsync-files';
+const DESTRUCTIVE_UPDATES_FLAG = 'allow-destructive-graphql-schema-updates';
+const PROVIDER_NAME = 'awscloudformation';
+
+type SanityCheckRules = {
+  diffRules: DiffRule[];
+  projectRules: ProjectRule[];
+};
 
 const warnOnAuth = (map: $TSObject, docLink: string): void => {
   const unAuthModelTypes = Object.keys(map).filter(type => !map[type].includes('auth') && map[type].includes('model'));
@@ -68,18 +74,15 @@ const warnOnAuth = (map: $TSObject, docLink: string): void => {
  * Transform GraphQL Schema
  */
 export const transformGraphQLSchemaV2 = async (context: $TSContext, options): Promise<DeploymentResources | undefined> => {
+  let resourceName: string;
   const backEndDir = pathManager.getBackendDirPath();
   const flags = context.parameters.options;
   if (flags['no-gql-override']) {
     return undefined;
   }
 
-  let { parameters } = options;
+  let { resourceDir, parameters } = options;
   const { forceCompile } = options;
-  const resourceDir = await contextUtil.getResourceDir(context, options);
-  if (!resourceDir) {
-    return undefined;
-  }
 
   // Compilation during the push step
   const { resourcesToBeCreated, resourcesToBeUpdated, allResources } = await context.amplify.getResourceStatus(AmplifyCategories.API);
@@ -101,6 +104,38 @@ export const transformGraphQLSchemaV2 = async (context: $TSContext, options): Pr
     resources = resources.concat(allResources);
   }
   resources = resources.filter(resource => resource.service === 'AppSync');
+
+  if (!resourceDir) {
+    // There can only be one appsync resource
+    if (resources.length > 0) {
+      const resource = resources[0];
+      if (resource.providerPlugin !== PROVIDER_NAME) {
+        return undefined;
+      }
+      const { category } = resource;
+      ({ resourceName } = resource);
+      resourceDir = path.normalize(path.join(backEndDir, category, resourceName));
+    } else {
+      // No appsync resource to update/add
+      return undefined;
+    }
+  }
+
+  let previouslyDeployedBackendDir = options.cloudBackendDirectory;
+  if (!previouslyDeployedBackendDir) {
+    if (resources.length > 0) {
+      const resource = resources[0];
+      if (resource.providerPlugin !== PROVIDER_NAME) {
+        return undefined;
+      }
+      const { category } = resource;
+      resourceName = resource.resourceName;
+      const cloudBackendRootDir = pathManager.getCurrentCloudBackendDirPath();
+      /* eslint-disable */
+      previouslyDeployedBackendDir = path.normalize(path.join(cloudBackendRootDir, category, resourceName));
+      /* eslint-enable */
+    }
+  }
 
   const parametersFilePath = path.join(resourceDir, PARAMETERS_FILENAME);
 
@@ -148,9 +183,29 @@ export const transformGraphQLSchemaV2 = async (context: $TSContext, options): Pr
     }
   }
 
+  // for auth transformer we get any admin roles and a cognito identity pool to check for
+  // potential authenticated roles outside of the provided authRole
+  const adminRoles = await getAdminRoles(context, resourceName);
+  const identityPoolId = await getIdentityPoolId(context);
+
+  // for the predictions directive get storage config
+  const s3Resource = s3ResourceAlreadyExists();
+  const storageConfig = s3Resource ? getBucketName(s3Resource) : undefined;
+
   const buildDir = path.normalize(path.join(resourceDir, 'build'));
   const schemaFilePath = path.normalize(path.join(resourceDir, SCHEMA_FILENAME));
   const schemaDirPath = path.normalize(path.join(resourceDir, SCHEMA_DIR_NAME));
+  let deploymentRootKey = await getPreviousDeploymentRootKey(previouslyDeployedBackendDir);
+  if (!deploymentRootKey) {
+    const deploymentSubKey = await CloudformationProviderFacade.hashDirectory(context, resourceDir);
+    deploymentRootKey = `${ROOT_APPSYNC_S3_KEY}/${deploymentSubKey}`;
+  }
+  const projectBucket = options.dryRun ? 'fake-bucket' : getProjectBucket();
+  const buildParameters = {
+    ...parameters,
+    S3DeploymentBucket: projectBucket,
+    S3DeploymentRootKey: deploymentRootKey,
+  };
 
   // If it is a dry run, don't create the build folder as it could make a follow-up command
   // to not to trigger a build, hence a corrupt deployment.
@@ -160,6 +215,9 @@ export const transformGraphQLSchemaV2 = async (context: $TSContext, options): Pr
 
   const project = await loadProject(resourceDir);
 
+  const lastDeployedProjectConfig = fs.existsSync(previouslyDeployedBackendDir)
+    ? await loadProject(previouslyDeployedBackendDir)
+    : undefined;
   const transformerVersion = await ApiCategoryFacade.getTransformerVersion(context);
   const docLink = getGraphQLTransformerAuthDocLink(transformerVersion);
   const sandboxModeEnabled = schemaHasSandboxModeEnabled(project.schema, docLink);
@@ -176,15 +234,65 @@ export const transformGraphQLSchemaV2 = async (context: $TSContext, options): Pr
 
   searchablePushChecks(context, directiveMap.types, parameters[ResourceConstants.PARAMETERS.AppSyncApiName]);
 
+  const transformerListFactory = await getTransformerFactory(context, resourceDir);
+
   if (sandboxModeEnabled && options.promptApiKeyCreation) {
     const apiKeyConfig = await showSandboxModePrompts(context);
     if (apiKeyConfig) authConfig.additionalAuthenticationProviders.push(apiKeyConfig);
   }
 
-  const buildConfig = await generateTransformerOptions(context, options);
-  if (!buildConfig) {
-    return undefined;
+  let searchableTransformerFlag = false;
+
+  if (directiveMap.directives.includes('searchable')) {
+    searchableTransformerFlag = true;
   }
+
+  // construct sanityCheckRules
+  const ff = new AmplifyCLIFeatureFlagAdapter();
+  const isNewAppSyncAPI: boolean = resourcesToBeCreated.some(resource => resource.service === 'AppSync');
+  const allowDestructiveUpdates = context?.input?.options?.[DESTRUCTIVE_UPDATES_FLAG] || context?.input?.options?.force;
+  const sanityCheckRules = getSanityCheckRules(isNewAppSyncAPI, ff, allowDestructiveUpdates);
+  let resolverConfig = {};
+  if (!_.isEmpty(resources)) {
+    resolverConfig = await context.amplify.invokePluginMethod(
+      context,
+      AmplifyCategories.API,
+      AmplifySupportedService.APPSYNC,
+      'getResolverConfig',
+      [context, resources[0].resourceName],
+    );
+  }
+
+  /**
+   * if Auth is not migrated , we need to fetch resolver Config from transformer.conf.json
+   * since above function will return empty object
+   */
+  if (_.isEmpty(resolverConfig)) {
+    resolverConfig = project.config.ResolverConfig;
+  }
+
+  const buildConfig: ProjectOptions<TransformerFactoryArgs> = {
+    ...options,
+    buildParameters,
+    projectDirectory: resourceDir,
+    transformersFactory: transformerListFactory,
+    transformersFactoryArgs: {
+      addSearchableTransformer: searchableTransformerFlag,
+      storageConfig,
+      authConfig,
+      adminRoles,
+      identityPoolId,
+    },
+    rootStackFileName: 'cloudformation-template.json',
+    currentCloudBackendDirectory: previouslyDeployedBackendDir,
+    minify: options.minify,
+    projectConfig: project,
+    lastDeployedProjectConfig,
+    authConfig,
+    sandboxModeEnabled,
+    sanityCheckRules,
+    resolverConfig,
+  };
 
   const transformerOutput = await buildAPIProject(context, buildConfig);
 
@@ -201,12 +309,109 @@ place .graphql files in a directory at ${schemaDirPath}`);
   return transformerOutput;
 };
 
+const getProjectBucket = (): string => {
+  const meta: $TSMeta = stateManager.getMeta(undefined, { throwIfNotExist: false });
+  const projectBucket = meta?.providers ? meta.providers[PROVIDER_NAME].DeploymentBucketName : '';
+  return projectBucket;
+};
+
+const getPreviousDeploymentRootKey = async (previouslyDeployedBackendDir: string): Promise<string|undefined> => {
+  // this is the function
+  let parameters;
+  try {
+    const parametersPath = path.join(previouslyDeployedBackendDir, `build/${PARAMETERS_FILENAME}`);
+    const parametersExists = fs.existsSync(parametersPath);
+    if (parametersExists) {
+      const parametersString = await fs.readFile(parametersPath);
+      parameters = JSON.parse(parametersString.toString());
+    }
+    return parameters.S3DeploymentRootKey;
+  } catch (err) {
+    return undefined;
+  }
+};
+
+/**
+ * Check if storage exists in the project if not return undefined
+ */
+const s3ResourceAlreadyExists = (): string | undefined => {
+  try {
+    let resourceName: string;
+    const amplifyMeta: $TSMeta = stateManager.getMeta(undefined, { throwIfNotExist: false });
+    if (amplifyMeta?.[AmplifyCategories.STORAGE]) {
+      const categoryResources = amplifyMeta[AmplifyCategories.STORAGE];
+      Object.keys(categoryResources).forEach(resource => {
+        if (categoryResources[resource].service === AmplifySupportedService.S3) {
+          resourceName = resource;
+        }
+      });
+    }
+    return resourceName;
+  } catch (error) {
+    if (error.name === 'UndeterminedEnvironmentError') {
+      return undefined;
+    }
+    throw error;
+  }
+};
+
+const getBucketName = (s3ResourceName: string): { bucketName: string } => {
+  const amplifyMeta = stateManager.getMeta();
+  const stackName = amplifyMeta.providers.awscloudformation.StackName;
+  const s3ResourcePath = pathManager.getResourceDirectoryPath(undefined, AmplifyCategories.STORAGE, s3ResourceName);
+  const cliInputsPath = path.join(s3ResourcePath, 'cli-inputs.json');
+  let bucketParameters: $TSObject;
+  // get bucketParameters 1st from cli-inputs , if not present, then parameters.json
+  if (fs.existsSync(cliInputsPath)) {
+    bucketParameters = JSONUtilities.readJson(cliInputsPath);
+  } else {
+    bucketParameters = stateManager.getResourceParametersJson(undefined, AmplifyCategories.STORAGE, s3ResourceName);
+  }
+  const bucketName = stackName.startsWith('amplify-')
+    ? `${bucketParameters.bucketName}\${hash}-\${env}`
+    : `${bucketParameters.bucketName}${s3ResourceName}-\${env}`;
+  return { bucketName };
+};
+
+type TransformerFactoryArgs = {
+  addSearchableTransformer: boolean;
+  authConfig: $TSAny;
+  storageConfig?: $TSAny;
+  adminRoles?: Array<string>;
+  identityPoolId?: string;
+};
+
+/**
+ * ProjectOptions Type Definition
+ */
+type ProjectOptions<T> = {
+  buildParameters: {
+    S3DeploymentBucket: string;
+    S3DeploymentRootKey: string;
+  };
+  projectDirectory: string;
+  transformersFactory: (options: T) => Promise<TransformerPluginProvider[]>;
+  transformersFactoryArgs: T;
+  rootStackFileName: 'cloudformation-template.json';
+  currentCloudBackendDirectory?: string;
+  minify: boolean;
+  lastDeployedProjectConfig?: TransformerProjectConfig;
+  projectConfig: TransformerProjectConfig;
+  resolverConfig?: ResolverConfig;
+  dryRun?: boolean;
+  authConfig?: AppSyncAuthConfiguration;
+  stacks: Record<string, Template>;
+  sandboxModeEnabled?: boolean;
+  sanityCheckRules: SanityCheckRules;
+  overrideConfig: OverrideConfig;
+};
+
 /**
  * buildAPIProject
  */
 const buildAPIProject = async (
   context: $TSContext,
-  opts: TransformerProjectOptions<TransformerFactoryArgs>,
+  opts: ProjectOptions<TransformerFactoryArgs>,
 ): Promise<DeploymentResources|undefined> => {
   const schema = opts.projectConfig.schema.toString();
   // Skip building the project if the schema is blank
@@ -236,7 +441,7 @@ const buildAPIProject = async (
   return builtProject;
 };
 
-const _buildProject = async (opts: TransformerProjectOptions<TransformerFactoryArgs>): Promise<DeploymentResources> => {
+const _buildProject = async (opts: ProjectOptions<TransformerFactoryArgs>): Promise<DeploymentResources> => {
   const userProjectConfig = opts.projectConfig;
   const stackMapping = userProjectConfig.config.StackMapping;
   const userDefinedSlots = {
