+++ packages/amplify-graphql-transformer-core/src/index.ts
@@ -1,13 +1,7 @@
 import { print } from 'graphql';
 import { EXTRA_DIRECTIVES_DOCUMENT } from './transformation/validation';
-
 export { GraphQLTransform, GraphQLTransformOptions, SyncUtils } from './transformation';
-export {
-  DeploymentResources,
-  UserDefinedSlot,
-  UserDefinedResolver,
-  OverrideConfig,
-} from './transformation/types';
+export { DeploymentResources, UserDefinedSlot, UserDefinedResolver } from './transformation/types';
 export { validateModelSchema } from './transformation/validation';
 export {
   ConflictDetectionType,
